Human action classification with still images using CNN
In this research the performance of all the traditional methods, deep learning methods, ResNet50 with SVM classification and CNN was tested on the task of human action recognition in still images. A custom dataset was created by the author for this research. The dataset consisted of images of 7 human activities, which are yawning, phoning, sitting, standing, walking, running and hand waving. Because of underfitting, conventional process models produced poor performance. The same issue arose with deep learning techniques. Due to underfitting, they also had very low accuracy. For feature extraction, ResNet50 was used because it produced the best results among deep learning methods, and SVM was used for classification because it produced the best results among all conventional methods, but the results were not satisfactory. Finally, in order to get a better result for the custom dataset, a custom CNN model was introduced, and the model produced promising results.